Added driver and README portion of power method.
