+++ test/power_A.dat
@@ -1,0 +1,2 @@
+2. 1.
+5. 7.
