+++ test/power_w.dat
@@ -1,0 +1,2 @@
+1.
+1.
