+++ test/power_x_0.dat
@@ -1,0 +1,2 @@
+1.
+1.
